Refactoring graph test
